Salary Prediction App
It's an application to predict the salary of the person who is currently or wants to be a software developer over the world. It can predict the salary of those people based on some parameters such as education and experience. I deployed  Machine learning algorithms such as linear regression and decision tree regression on the dataset and also perform some cleaning duties on a dataset for effective usages.